New macros and field formats for proc-as-article
The unified style sheet suggests that conference proceedings and working
papers with ISSNs should be treated as types of the article entry.

The current implementation of
biblatex-sp-unified (512f11657199a6044f7663da454f3eac338bdbd5) handles
this by requiring users to enter these entries as @article's in the
database.

Rather than having to do this at the databse level, these new macros and
field format will allow for writing a new driver that can do this at the
level of the style file.