Remove gratuitous \newunit and \newunit\newblock
Bugs were introduced in cb3feb9389cb58e4b579048dd8125a192997babc and
86841da18af5ec38745b268f187e402e61102fcc where \newunit and
\newunit\newblock were being issued in places that they should not be
issued. \newunit inserts \newunitpunct into the punctuation
buffer (usually a period), but we do not want to have a period here
because conference proceedings with ISSNs ought to typeset as articles,
according to the Unified Style Sheet, and articles do not have periods
in between the journal title and the volume(number) information.

This commit fixes #6.