Added DOIs to the test bibliography
For testing the display of DOIs, three DOIs were added to the test bibliography. To match the output shown in the original Unified Stylesheet document, the test `.tex` file sets the display of DOIs to false. Usually, the `unified` style will in fact display DOIs. We follow the Crossref display guidelines, but the user can choose to load biblatex with the `compactdois` package option to display DOIs more compactly.