Treat `pubstate` field as label fallback
biblatex doesn't like things like "forthcoming'' in the year/date field. The canonical way to do this is to put "forthcoming" in the `pubstate` field. Since we want that to be treated as the label, we need to adjust some things. What we do here is suggested at https://tex.stackexchange.com/a/444414/70432.